+++ SpaceShip.cpp
@@ -9,7 +9,7 @@ SpaceShip::SpaceShip()
 		0.0f, 0.0f, 0.03f, 0.0f,
 		0.0f, 0.0f, 100.0f, 1.0f
 	};
-	pi = atan(1) * 4;
+	pi = static_cast<float>(atan(1) * 4);
 	this->m_rotationAngle = pi;
 }
